@@ -10,7 +10,7 @@ var (
 		Lat: 37,
 		Lng: -122,
 	}
-	cell     = LatLngToCell(geo, 15)
+	cell, _  = LatLngToCell(geo, 15)
 	addr     = cell.String()
 	geoBndry CellBoundary
 	cells    []Cell
@@ -31,30 +31,30 @@ func BenchmarkFromString(b *testing.B) {
 
 func BenchmarkToGeoRes15(b *testing.B) {
 	for n := 0; n < b.N; n++ {
-		geo = CellToLatLng(cell)
+		geo, _ = CellToLatLng(cell)
 	}
 }
 
 func BenchmarkFromGeoRes15(b *testing.B) {
 	for n := 0; n < b.N; n++ {
-		cell = LatLngToCell(geo, 15)
+		cell, _ = LatLngToCell(geo, 15)
 	}
 }
 
 func BenchmarkToGeoBndryRes15(b *testing.B) {
 	for n := 0; n < b.N; n++ {
-		geoBndry = CellToBoundary(cell)
+		geoBndry, _ = CellToBoundary(cell)
 	}
 }
 
 func BenchmarkHexRange(b *testing.B) {
 	for n := 0; n < b.N; n++ {
-		cells = cell.GridDisk(10)
+		cells, _ = cell.GridDisk(10)
 	}
 }
 
 func BenchmarkPolyfill(b *testing.B) {
 	for n := 0; n < b.N; n++ {
-		cells = PolygonToCells(validGeoPolygonHoles, 15)
+		cells, _ = PolygonToCells(validGeoPolygonHoles, 15)
 	}
 }
